Revamp and simplification of the sass2json logic and approach. Works with dart-sass currently.
