+++ .prettierignore
@@ -12,4 +12,4 @@ package-lock.json
 
 _icons-generated.scss
 
-scssVariables.json
+sass-variables.json
