+++ apps/node-pl/pattern-lab/_data/sass-variables.json
@@ -1,0 +1,1 @@
+{"scssColors":{"primary":"rgba(0,123,2551)","secondary":"rgba(108,117,1251)","success":"rgba(40,167,691)","info":"rgba(23,162,1841)","warning":"rgba(255,193,71)","danger":"rgba(220,53,691)","light":"rgba(244,244,2441)","dark":"rgba(43,43,431)","pink":"rgba(255,192,2031)","goldenrod":"rgba(218,165,321)"},"primaryColors":{"pink":"rgba(255,192,2031)","goldenrod":"rgba(218,165,321)"},"neutralHues":{"lightest-gray":"rgba(244,244,2441)","light-gray":"rgba(228,228,2281)","gray":"rgba(213,213,2131)","dark-gray":"rgba(182,182,1821)","darkest-gray":"rgba(43,43,431)"},"scssBreakpoints":{"xs":0,"sm":"576px","md":"768px","lg":"992px","xl":"1200px"},"scssSpacing":{"0":0,"1":"0.25rem","2":"0.5rem","3":"1rem","4":"1.5rem","5":"3rem","6":"4rem"},"fontFamilies":{"$font-family-sans-serif":"Open Sans, Helvetica Neue, Helvetica, Arial, sans-serif","$font-family-serif":"Oswald, Times New Roman, serif","$font-family-monospace":"Source Code Pro, Courier New, monospace","$headings-font-family":"Oswald, Times New Roman, serif"},"fontSizes":{"base":"1rem","lg":"1.25rem","sm":"0.875rem"},"fontWeights":{"light":300,"normal":400,"bold":700},"scssIcons":["chevron-down","chevron-left","chevron-right","chevron-up","close","menu","person","search"]}
