--- apps/node-pl/pattern-lab/_data/scssVariables.json
@@ -1,68 +1,0 @@
-{
-  "scssColors": {
-    "primary": "rgba(0, 123, 255, 1)",
-    "secondary": "rgba(108, 117, 125, 1)",
-    "success": "rgba(40, 167, 69, 1)",
-    "info": "rgba(23, 162, 184, 1)",
-    "warning": "rgba(255, 193, 7, 1)",
-    "danger": "rgba(220, 53, 69, 1)",
-    "light": "rgba(244, 244, 244, 1)",
-    "dark": "rgba(43, 43, 43, 1)",
-    "pink": "rgba(255, 192, 203, 1)",
-    "goldenrod": "rgba(218, 165, 32, 1)"
-  },
-  "primaryColors": {
-    "pink": "rgba(255, 192, 203, 1)",
-    "goldenrod": "rgba(218, 165, 32, 1)",
-    "derpcolor": "rgba(255, 255, 222, 0.7)"
-  },
-  "neutralHues": {
-    "lightest-gray": "rgba(244, 244, 244, 1)",
-    "light-gray": "rgba(228, 228, 228, 1)",
-    "gray": "rgba(213, 213, 213, 1)",
-    "dark-gray": "rgba(182, 182, 182, 1)",
-    "darkest-gray": "rgba(43, 43, 43, 1)"
-  },
-  "scssBreakpoints": {
-    "xs": 0,
-    "sm": "576px",
-    "md": "768px",
-    "lg": "992px",
-    "xl": "1200px"
-  },
-  "scssSpacing": {
-    "0": 0,
-    "1": "0.25rem",
-    "2": "0.5rem",
-    "3": "1rem",
-    "4": "1.5rem",
-    "5": "3rem",
-    "6": "4rem"
-  },
-  "fontFamilies": {
-    "$font-family-sans-serif": "Open Sans, Helvetica Neue, Helvetica, Arial, sans-serif",
-    "$font-family-serif": "Oswald, Times New Roman, serif",
-    "$font-family-monospace": "Source Code Pro, Courier New, monospace",
-    "$headings-font-family": "Oswald, Times New Roman, serif"
-  },
-  "fontSizes": {
-    "base": "1rem",
-    "lg": "1.25rem",
-    "sm": "0.875rem"
-  },
-  "fontWeights": {
-    "light": 300,
-    "normal": 400,
-    "bold": 700
-  },
-  "scssIcons": [
-    "chevron-down",
-    "chevron-left",
-    "chevron-right",
-    "chevron-up",
-    "close",
-    "menu",
-    "person",
-    "search"
-  ]
-}
