+++ apps/node-pl/pattern-lab/_patterns/00-protons-demo/spacing.twig
@@ -4,7 +4,7 @@
  * Documenting Spacing Examples from Bootstrap.
  *
  * Available variables:
- * - scssSpacing: [array] Available in scssVariables.
+ * - scssSpacing: [array] Available in sass-variables.json.
  *
  * @see https://getbootstrap.com/docs/4.0/utilities/spacing/
  */
@@ -15,10 +15,23 @@
     <div class="col">
       <h1 class="mb-3">Spacing</h1>
       <div class="mb-3 p-3 bg-light border border-secondary rounded">
-        <p class="lead">To add to these items, adjust <code>spacers()</code> in <code>non-printing/_bootstrap-overrides.scss</code>.</p>
-        <p><code>spacers(0)</code> produces a value of <code>0</code> and is allowed but not shown in this demo.</p>
-        <p>The math for rem calculations here is based on a browser value of 16px as a <code>$spacer</code> value. To find out your spacer in pixels, simply multiply the rem value in the spacers list by this spacer value. For example, a 1.5rem value in <code>$spacers</code> is equivalent to 24px.</p>
-        <div><a href="https://getbootstrap.com/docs/4.0/utilities/spacing/" target="_blank" rel="noopener noreferrer">Bootstrap Documentation</a></div>
+        <p class="lead">To add to these items, adjust
+          <code>spacers()</code>
+          in
+          <code>non-printing/_bootstrap-overrides.scss</code>.</p>
+        <p>
+          <code>spacers(0)</code>
+          produces a value of
+          <code>0</code>
+          and is allowed but not shown in this demo.</p>
+        <p>The math for rem calculations here is based on a browser value of 16px as a
+          <code>$spacer</code>
+          value. To find out your spacer in pixels, simply multiply the rem value in the spacers list by this spacer value. For example, a 1.5rem value in
+          <code>$spacers</code>
+          is equivalent to 24px.</p>
+        <div>
+          <a href="https://getbootstrap.com/docs/4.0/utilities/spacing/" target="_blank" rel="noopener noreferrer">Bootstrap Documentation</a>
+        </div>
       </div>
       <hr>
 
@@ -26,7 +39,10 @@
         {% if spacer != 0 %}
           <div style="margin: 10px;">
             <div style="height: {{ value }}; width: {{ value }}; outline: solid 1px black; display: inline-block; margin-right: 5px;"></div>
-            <span>{{ value }} : <code>spacers({{ spacer }})</code></span>
+            <span>{{ value }}
+              :
+              <code>spacers({{ spacer }})</code>
+            </span>
           </div>
         {% endif %}
       {% endfor %}
