+++ apps/node-pl/scss/_scss2json.scss
@@ -31,11 +31,11 @@ $spacing-demo: $spacers;
 // Import your Customized Variables here, example:
 //@import '../_patterns/01-atoms/button/_button.scss';
 
-// Exports Data to source/_data/scssVariables.json
+// Exports Data to source/_data/sass-variables.json
 // This json file populates demo twig patterns related to config items.
 // Include your Customized Sass Maps here to JSONIFY
 @include export-data(
-  'scssVariables.json',
+  'sass-variables.json',
   (
     scssColors: $theme-colors,
     primaryColors: $primary-colors,
