+++ apps/node-pl/webpack.config.js
@@ -6,7 +6,7 @@
 const path = require('path');
 
 const { DefinePlugin } = require('webpack');
-const sassExportData = require('@theme-tools/sass-export-data');
+// const sassExportData = require('@theme-tools/sass-export-data');
 
 // Plugins
 const RunScriptOnFiletypeChange = require('../../tools/webpack/run-script-on-filetype-change');
@@ -41,11 +41,9 @@ const shared = {
               sassOptions: {
                 // Used to generate JSON about variables like colors, fonts
                 functions: {
-                  'export_data($file, $value)': sassExportData(path.resolve(APP_PATH, 'pattern-lab/_data/')),
-                  // ...sassExportData({
-                  //   name: 'export_data',
-                  //   path: path.resolve(APP_PATH, 'pattern-lab/_data/'),
-                  // }),
+                  'export_data($file, $value)': sassExportData(
+                    path.resolve(APP_PATH, 'pattern-lab/_data/')
+                  ),
                 },
               },
             },
