+++ package.json
@@ -82,9 +82,11 @@
     "eslint-plugin-vue": "^5.2.3",
     "fibers": "^4.0.2",
     "file-loader": "^4.2.0",
+    "fs-extra": "^8.1.0",
     "jest": "^24.9.0",
     "jest-fetch-mock": "^2.1.2",
     "jest-vue-preprocessor": "^1.7.0",
+    "jsondiffpatch": "^0.3.11",
     "mini-css-extract-plugin": "^0.8.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
